// frontend/app/controller/model/workflow/stepDTO.ts
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { BaseDto } from 'app/zynerator/dto/BaseDto.model';
import { DocumentDto } from '../Document.model';
import { CommentaireDTO } from './commentaireDTO';

import { StepPresetDTO } from './stepPresetDTO';
import { WorkflowDTO } from './workflowDTO';

export interface StepDTO extends BaseDto { 
   
    stepPreset: StepPresetDTO;
    status: StepDTO.StatusEnum;
    documents?: Array<DocumentDto>;
    documentsActions?: Array<DocumentDto>;
    workflowId: number;
    discussions?: Array<CommentaireDTO>;
    actions?: ACTION[];
    validateDate?: string;

    createdOn: string;
    updatedOn: string;
    createdBy: string;
    updatedBy: string;

}
export namespace StepDTO {
    export type StatusEnum = 'WAITING' | 'IN_PROGRESS'| 'REJECTED' | 'DONE' | 'Annulled' | 'PARTIAL' | 'COMPLEMENT';
    export const StatusEnum = {
        WAITING: 'WAITING' as StatusEnum,
        INPROGRESS: 'IN_PROGRESS' as StatusEnum,
        REJECTED: 'REJECTED' as StatusEnum,
        DONE: 'DONE' as StatusEnum,
        Annulled: 'Annulled' as StatusEnum,
        PARTIAL: 'PARTIAL' as StatusEnum,
        COMPLEMENT: 'COMPLEMENT' as StatusEnum

    };
}
export enum ACTION {
    SIGN = 'SIGN',
    APPROVE = 'APPROVE',
    REJECT = 'REJECT',
    PARAPHER = 'PARAPHER',
    PRESIGNER = 'PRESIGNER',
    ENVOI_COURRIER = 'ENVOI_COURRIER',
  }